+++ rocAL/rocAL/include/loaders/image/node_fused_jpeg_crop_single_shard.h
@@ -22,7 +22,7 @@ public:
     /// for example if there are 10 images in the dataset and load_batch_count is 3, the loader repeats 2 images as if there are 12 images available.
     void init(unsigned shard_id, unsigned shard_count, const std::string &source_path, const std::string &json_path, StorageType storage_type,
               DecoderType decoder_type, bool shuffle, bool loop, size_t load_batch_count, RocalMemType mem_type, std::shared_ptr<MetaDataReader> meta_data_reader,
-              unsigned num_attempts, std::vector<double> &random_area, std::vector<double> &random_aspect_ratio);
+              unsigned num_attempts, std::vector<float> &random_area, std::vector<float> &random_aspect_ratio);
 
     std::shared_ptr<LoaderModule> get_loader_module();
 protected:
@@ -30,6 +30,6 @@ protected:
     void update_node() override {};
 private:
     std::shared_ptr<ImageLoader> _loader_module = nullptr;
-    std::vector<double> _random_area, _random_aspect_ratio;
+    std::vector<float> _random_area, _random_aspect_ratio;
     unsigned num_attempts;
 };
